4.4.1r15 - hbbtv video player fixes - fixed possible Alsa Output crash - eListbox: don't react when attributes are re-set to current value - ignore hdmitx hpd events when in idle mode

+++ usr/include/enigma2/lib/hbbtv/hbbtv.h
@@ -109,8 +109,8 @@ public:
 	void onCurrentServiceStop();
 	void onCurrentServiceEvent(iPlayableService *playableService, int what);
 	ePtr<iPlayableService> getPlayableService();
-	long getPlayTime();
-	long getPlayPosition();
+	int getPlayTime();
+	int getPlayPosition();
 	void setVolume(int volume);
 	const std::list<eServiceReference> getServiceList();
 	ePtr<eServiceEvent> getEvent(int nownext);
